Simplify dtype and device settings

+++ gllm/model_runner.py
@@ -39,7 +39,7 @@ class ModelRunner():
         memory_manager_cls = PrefixMemoryManager if self.enable_prefix_caching else MemoryManager
         self.memory_manager = memory_manager_cls(
             gpu_memory_util=self.gpu_memory_util, num_layers=self.model.num_layers,
-            dtype=self.model.dtype, page_size=self.page_size, kv_head_num=self.model.num_kv_heads,
+            dtype=self.model_loader.dtype, page_size=self.page_size, kv_head_num=self.model.num_kv_heads,
             kv_head_dim=self.model.head_dim, vocab_size=self.model_loader.vocab_size)
 
     def encode(self, content, chat: bool = False):
